GameStates
This is the full list of files I created for the Games States. It works
as far as I can tell and be easy to edit.

Notes - Using placeholder art (and colours) to show that something is
actually happening. The buttons currently are all "Start Game" but its
just the first one I found on Google.
Some of the objects have code attached, where as some use the inbuilt
gamemaker actions, this is due to the mythical map, I thought it was
better if when entering/exiting the character menu, it remembered what
version of the map you were using.

Any questoins just ask