Ell:1.5 - Editor - Locations stored and loaded
Right, I have created a new version, specifically, this is the editor.
It uses the current (now old) GameStates (client) code but with the
added functionality of dynamically storing/reading locations. I have
commented heavily the new code (under editor in objects, all important
changes are done here). I need to do a little work to apply this
functionality to the client (GameStates) version, but it will be a case
of copying the reading of an ini file over, will take about 10 mins, I
will do this later.