1.7.9 - Editor - Various (1.8.0 preupdate)
Location add screen has generally been improved. I have adapted the text
boxes so that they can display non editable text and I have reorganised
the screen itself to create more room.
You can now change a location on the locationAdd screen, there is a
button, it will auto update the ini file and be visible on both maps or
either map.
I have started to add events, if you click, add event, this will add a
new blank event (all info reads 0), I will set this up so that the user
can edit these events. Currently events have a type, week and mythical
status.
Type - the event type, this will be 0 for no event, 1 onwards for the
different event types. Events will have a seperate event file, that will
be editable on another screen (accessible from this or other screens).
Week and mythical are fairly self explanitory, the week will be used to
further limit the events on the map screen. The mythical status will
display if the event is on a mythical map screen or normal. I need to
limit this further so that if the location is only visible on mythical,
the event can only be mythical, on normal, etc.
The folder has not been updated, this is so I can track the changes (if
I update the folder, it will update every file and it is useless for
trackign changes) for when I update the client later today.