Detection changes and variable code added for "times"
I have added a "Detection" for event objects, this allows us to know
when a location is clicked and should be somewhat recipricle for when an
event is clicked.

This means i can create a sudo timer for days and times in game, so lets
say the day begins at 9 am and every click increments this time by say,
two hours? lets say the the day ends at 11 pm, that means the play can
do 6 events in a day before they are forced to go to day 2.

ofcourse this can be changed to suit what ever need, so i could make it
so the time incremented by each even is different allowing for a
different amount and combination of events to happen each day.