Ell: V1.8 - Editor and Client
Client now updated to include the same code as the editor (see previous
update).
I have also added different buildings, so if you create a location, then
change "spriteRef" inside the ini file, to 0-1-2, you can swap the
building type, it will automatically update in the game, I have set this
up so that the buildings icons can be changed.