Ell:1.8.9 - Editor - Events/changes
I have started adding events, you can now add an event, see its info and
cycle through all events (if there is more than one). Mythical/Normal
status and cycle now only effect the event/relect the events status. So,
once an event is created, you can cycle through the mythical status' as
before, this will then update in the events mythical status, which in
turn effects the locations mythical status. This is so that it
automatically sets the map view based on the events in that location.
Eventually, individual events will have their own set status (that can
be edited inside the event itself), the info will then just be for user
convenience.
The "events type" will eventually become its name, for us however, this
will be used as the key/id for the event.
I have renamed all the objects (I will do sprites for full 1.9 release)
to a better naming system, they are a bit easier to see, I have changed
the filtering on the list too to make it easier to see.
You can now also use the arrows in the corners to cycle through all
Locations.
As before, I will update the client and then update the folder structure
to the next version (1.9).