Ell:1.9 - Client and Editor
Client and editor up to date.
All files renamed to a better naming system.
The editor can now add events, I need to set this up a bit more and add
cards before I can actually add the event handling to the client. But it
is getting there, the next patch should be quite big and I am hoping to
have the ability to add events, cards, and have some action from this by
then. Once that is done it will be a case of adding the player stats to
a file and accessing these on startup for use within the game. Then
saving that when the user does anything.