Ell:2.2- Editor - Week
Now tracks week, it will automatically filter locations out based on the
events week stored inside the locations.
This is needed before I can move changes over to client. Rather than
rewriting the client and then having to add to the editor, I decided to
add to the editor then copy that over (means I only have to play about
with the actual code, once).
The client is VERY behind at the moment, so it will take me a little
while to bring it up to the editor. This my next task.