Ell:2.2 - Editor - Various
I have changed a couple of things around on the editor, this has been to
help with the transfer over to the client.
There can only be one event per week, per location now. This is becuase
before, you could have multiple events at a location on the same week,
but how does this work for the client? It seems stupid to go to a
location, select the event from a list, then select a card from a
list..... if there is an event at a location, there should be one event,
it should show the list of cards, that contain "lecture", "bunk", etc.
Currently, it just reverts back to what was entered before (so if you
have a week 1 and week 2, select both to week 1, the week to will revert
to week 2).
I have made it so that cards have a mythical status and the card select
on events filters it down to the same status. It didn't make sence to me
that, for example, you could select a mythical event and choose "Attend
lecture" card. The general card list has not been effected.