Ell:2.2.4 - Editor - Change/fix adding action cards
When adding action cards to an event, it wasnt correctly adding to the
total number (in the ini file anyway).
Also changed it so that it only creates a new event, if the current
event is blank (so you cna only have one blank action card at a time,
previously, if you changed an action card, it would create a new blank
one).