Ell:V3.1 - Client and Editor - Minor fix/change
I have fixed the client so it correctly shows locations on the map,
previosuly, it wasn't showing mythical/normal events correctly.
As a result of this fix, I have completely removed the mythical map
button and title. Now there is ONLY editor mode.
I have also updated the code that shows a different colour icon
depending on the mythical status, I feel this a much better way of
displaying that information.