Ell:V3.1 - Editor - Image Selection
I have made it so that on the location edit/add screen, you can select
the building sprite, currently there are only 2 sprites
(buildingspritenormal and buildingspritenormal1, so it only appears to
work on normal map buildings, but it is working on the others, just
nothing is there to change to).

This can easily be replicated for card selection, the code itself is
fairly simple.

I have also updated the way in which the building sprites are selected
on the map (in line with how I am now doing it on the locationAdd
screen), this hopefully has made it a lot more reliable when you switch
the events out, testing will be required to see if this is 100% though.