Ell:V1.23 - Editor - Location add fix + button position moved
Locations can now be added on when no locations already exist, this was
bugged due to me adding the building sprite to the locations in a recent
update. This is now fixed.
Moved the position of the "Select Event" button, depending on the length
of the text in fields to the left, there was overlap, now moved below to
recify this ever being an issue.