Ell:V4.1 - Editor - Map screen changes
Added text under each location, changed the buttons on the map screen to
be a list down the right hand side and I changed the way that instances
are checked/stored.

I have added arrays for a few things, building and text box instances
and the different cards, this makes it easier to reference them when
accessing elsewhere.