Added Latest Map/Menu Art and Zoom in/out
ok so i added the the most recent version of the main menu art and also
the latest map to the game. things worth noting, and also removed the
buttons on the screen and added a generic "move to map" object which
moves the player from the main screen to the mpa when they click and or
tap anywhere on the screen.

I started working on zooming in/out while on the map screen, the
functionality of it works although it conflicts and essentially
overwrites the drag to move function we already have within the game,
this isnt ideal but i have a good idea on how to make this work, for now
i have disabled the zoom in/out code although it is still there, i will
work on it at a later date

Changelog
-Removed Menu Buttons
-Added Object to menu screen allowing the rooms to change from menu to
map
-Added Main Menu Art work
-Added latest Map Artwork
-Created Object under "map" name Zoom
-Added create event to zoom which creates a variable called zoom, set it
to 1
-Created a step event in Zoom, takes screen width and height and * it by
Zoom
-added zoom controls to the step event, W zooming in, S zooming out