Ell:V4.5 - Editor and Client - Zoom
Client and editor now support zooming (mouse wheel up/down), this needs
testing on andriod to see how it will work, but at least this is another
step in the right direction.
Small quality of life change, any screen that has the back button on,
now supports right clicking to move back a screen.
Changed the way that some text is displayed on the map (week number is
now on the GUI).