Ell:V5.0 - Editor - Card images added
Can now add and select the sprites for each card.
Added a save button for when adding a new card, back will now go back
without saving unless save is clicked, not sure if I like this or not,
delete is there as a cancel so..... Feedback required really.
Also tidied up some of the text display on the map (anything in the 10s
and above for the number of card caused it to become unreadable, now
fixed).