Ell:V5.5 - Editor - Minor
When there are more than 20 cards, you can now scroll through the list
of cards.. this will be needed (or something similar for the client....
I am currently in the process of designing a card layout system where
you can view the cards on a page, this will be used in multiple places
within the client so is required. Seeing as the editor is already
working, I am using this as a testbed.