Ell:V5.6 - Editor - Card Lists
I have updated the card lists to show the actual cards on each card list
page, by default it will show the standard list but you can use the
button at the top of the page to swap to the card display (and back).

There are left/right buttons that will display depending on the list;
On standard list: The left most ones add/subtract 1 to the position,
moving the lits.
The right most arrows will move it to the bottom/top of the list.

On card display: The left arrows work the same as before.
The right arrows now move 4 spaces either way.