V5.6 - Editor and V4.1 - Client
Editor - Now has plus/minus button for int,char and well values so that
they can be set to plus or minus (text entry will not work for those
keys).
Client - Can now go through a location, selecting each card at each step
(except char cards). This is setup in the editor and needs to be copied
over, I have not uploaded the files to the web.
Also on first load you are required to enter your name, this will only
happen the first time, this is what creates your character info that is
edited in game and where the tutorial will begin.