Various Changes
Client:
- Cleaned up the card interaction, now has an info link button that will
take the user to the info card, rather than it being a seperate screen
as you go. I feel this looks a lot better.
- Fixed text entry on mobile, this wasn't working at all.
- Fixed downloading of ini files and sprites, there was an issue where
it wouldn't download the files correctly the first time, this was due to
the download for a required file not completing until the page has
finished, this was required to downloading sprites and obviously if
mising, it cannot download anything.
- Back button on Andriod (hardware) now supported.
Editor:
- Updated update functionality when entering information into text
fields, will now save the info as it goes, this was causing some
information to be lost if not used corerctly and generally a lot better
now.
- Removed some unecessary buttons

Both
- Added Info card that gives more room for info and removed the sprite
from the page, will remove the ability to add and view them elsewhere
shortly.